Add this year to copyright list.
Apparently this needs to be done in such a way that the range only includes
the years during which copyrightable changes were made. See
https://softwareengineering.stackexchange.com/a/210491.

Note that the change made in the prior commit may seem a trivial buildsystem
change. However, the whole point of this repo is demonstrating that exact
buildsystem change. So the change is material.